Remove old macros
