+++ rust/hyperlane-base/src/agent.rs
@@ -1,6 +1,4 @@
-use std::env;
-use std::fmt::Debug;
-use std::sync::Arc;
+use std::{env, fmt::Debug, sync::Arc};
 
 use async_trait::async_trait;
 use eyre::{Report, Result};
@@ -8,7 +6,7 @@ use futures_util::future::select_all;
 use tokio::task::JoinHandle;
 use tracing::{debug_span, instrument::Instrumented, Instrument};
 
-use crate::{cancel_task, metrics::CoreMetrics, settings::Settings};
+use crate::{metrics::CoreMetrics, settings::Settings};
 
 /// Properties shared across all hyperlane agents
 #[derive(Debug)]
@@ -86,7 +84,9 @@ pub fn run_all(
         let (res, _, remaining) = select_all(tasks).await;
 
         for task in remaining.into_iter() {
-            cancel_task!(task);
+            let t = task.into_inner();
+            t.abort();
+            t.await;
         }
 
         res?
