+++ rust/hyperlane-base/src/settings/mod.rs
@@ -72,9 +72,24 @@
 //! 5. Arguments passed to the agent on the command line.
 //!    E.g. `--originChainName ethereum`
 
+use std::fmt::Debug;
+
 pub use base::*;
-pub use chains::{ChainConf, ChainConnectionConf, CoreContractAddresses};
-pub use signers::{RawSignerConf, SignerConf};
+pub use chains::*;
+pub use checkpoint_syncer::*;
+use hyperlane_core::config::*;
+/// Export this so they don't need to import paste.
+#[doc(hidden)]
+pub use paste;
+use serde::Deserialize;
+pub use signers::*;
+pub use trace::*;
+
+mod envs {
+    pub use hyperlane_ethereum as h_eth;
+    pub use hyperlane_fuel as h_fuel;
+    pub use hyperlane_sealevel as h_sealevel;
+}
 
 /// AWS Credentials provider.
 pub(crate) mod aws_credentials;
@@ -86,3 +101,116 @@ pub(crate) mod loader;
 mod signers;
 /// Tracing subscriber management
 pub mod trace;
+#[macro_export]
+/// Declare a new settings block
+///
+/// This macro declares a settings struct for an agent. The new settings block
+/// contains a [`crate::Settings`] and any other specified attributes.
+///
+/// Please note that integers must be specified as `String` in order to allow
+/// them to be configured via env var. They must then be parsed in the
+/// [`Agent::from_settings`](crate::agent::Agent::from_settings)
+/// method.
+///
+/// ### Usage
+///
+/// ```ignore
+/// decl_settings!(Validator {
+///    validator: SignerConf,
+///    checkpointsyncer: CheckpointSyncerConf,
+///    reorgperiod: String,
+///    interval: String,
+/// });
+/// ```
+macro_rules! decl_settings {
+    (
+        $name:ident,
+        Parsed {
+            $($(#[$parsed_tags:meta])* $parsed_prop:ident: $parsed_type:ty,)*
+        },
+        Raw {
+            $($(#[$raw_tags:meta])* $raw_prop:ident: $raw_type:ty,)*
+        }$(,)?
+    ) => {
+        hyperlane_base::settings::paste::paste! {
+            #[doc = "Settings for `" $name "`"]
+            #[derive(Debug)]
+            pub struct [<$name Settings>] {
+                base: hyperlane_base::settings::Settings,
+                $(
+                    $(#[$parsed_tags])*
+                    pub(crate) $parsed_prop: $parsed_type,
+                )*
+            }
+
+            #[doc = "Raw settings for `" $name "`"]
+            #[derive(Debug, serde::Deserialize)]
+            #[serde(rename_all = "camelCase")]
+            pub struct [<Raw $name Settings>] {
+                #[serde(flatten, default)]
+                base: hyperlane_base::settings::deprecated_parser::DeprecatedRawSettings,
+                $(
+                    $(#[$raw_tags])*
+                    $raw_prop: $raw_type,
+                )*
+            }
+
+            impl AsMut<hyperlane_base::settings::deprecated_parser::DeprecatedRawSettings> for [<Raw $name Settings>] {
+                fn as_mut(&mut self) -> &mut hyperlane_base::settings::deprecated_parser::DeprecatedRawSettings {
+                    &mut self.base
+                }
+            }
+
+            // ensure the settings struct implements `FromRawConf`
+            const _: fn() = || {
+                fn assert_impl<T>()
+                where
+                    T: ?Sized + hyperlane_core::config::FromRawConf<[<Raw $name Settings>]>
+                {}
+
+                assert_impl::<[<$name Settings>]>();
+            };
+
+            impl std::ops::Deref for [<$name Settings>] {
+                type Target = hyperlane_base::settings::Settings;
+
+                fn deref(&self) -> &Self::Target {
+                    &self.base
+                }
+            }
+
+            impl AsRef<hyperlane_base::settings::Settings> for [<$name Settings>] {
+                fn as_ref(&self) -> &hyperlane_base::settings::Settings {
+                    &self.base
+                }
+            }
+
+            impl AsMut<hyperlane_base::settings::Settings> for [<$name Settings>] {
+                fn as_mut(&mut self) -> &mut hyperlane_base::settings::Settings {
+                    &mut self.base
+                }
+            }
+
+            impl hyperlane_base::NewFromSettings<> for [<$name Settings>] {
+                /// See `load_settings_object` for more information about how settings are loaded.
+                fn new() -> hyperlane_core::config::ConfigResult<Self> {
+                    hyperlane_base::settings::_new_settings::<[<Raw $name Settings>], [<$name Settings>]>(stringify!($name))
+                }
+            }
+        }
+    };
+}
+
+/// Static logic called by the decl_settings! macro. Do not call directly!
+#[doc(hidden)]
+pub fn _new_settings<'de, T, R>(name: &str) -> ConfigResult<R>
+where
+    T: Deserialize<'de> + AsMut<deprecated_parser::DeprecatedRawSettings> + Debug,
+    R: FromRawConf<T>,
+{
+    use crate::settings::loader::load_settings_object;
+    let root_path = ConfigPath::default();
+    let raw =
+        load_settings_object::<T, &str>(name, &[]).into_config_result(|| root_path.clone())?;
+    raw.parse_config(&root_path)
+}
